bgscan simple: Skip immediate scan on initial signal event
The driver is likely to indicate an immediate signal event when the
threshold value is configured. Since we do this immediately after
association, there is not much point in requesting a new scan to be
started based on this event.
(cherry picked from commit 1e6ef6455cd4c4213c269fa843ab53d9b392baf4)